convert measure_time() to output in seconds using double
measure_time() currently prints and returns some cpu tick count. For
identifying spots which take a long time is sufficient to relatively
compare cpu ticks. In order to get some real world reference we
convert the output to seconds. This helps when trying to match
to the waited real time.

+++ compat/mingw.h
@@ -459,7 +459,7 @@ extern int err_win_to_posix(DWORD winerr);
 extern const char *get_windows_home_directory();
 #define get_home_directory() get_windows_home_directory()
 
-extern int measure_time_aux(const char *file, int lineno, const char *fmt, ...);
+extern double measure_time_aux(const char *file, int lineno, const char *fmt, ...);
 #ifdef __GNUC__
 #define measure_time(...) \
 	measure_time_aux(__FILE__, __LINE__, __VA_ARGS__)
